[EASY] Move OpsHandler and MockHandler to their own file
[ghstack-poisoned]
